Adding Production-Ready Features and Rating Support

+++ src/main/java/com/bhawna/project/uber/uberApp/services/RiderService.java
@@ -15,7 +15,7 @@ public interface RiderService {
 
     RideDto cancelRide(Long rideId);
 
-    DriverDto rateDriver(Long rideId, Double rating);
+    DriverDto rateDriver(Long rideId, Integer rating);
 
     RiderDto getMyProfile();
